+++ src/CloseDialogHandle/index.d.ts
@@ -1,0 +1,7 @@
+import type { JSX, ParentProps } from 'solid-js'
+export function CloseDialogHandle($p:ParentProps<CloseDialogHandle__props_T>):JSX.Element
+export interface CloseDialogHandle__props_T {
+	class?:string
+	tabindex?:string
+	onclick?:(evt:MouseEvent)=>void
+}