+++ src/index.jsx
@@ -1,5 +1,6 @@
+export * from './CloseDialogHandle/index.jsx'
 export * from './ctx__Context/index.js'
-export * from './Matcha/index.js'
+export * from './Matcha/index.jsx'
 export * from './minifycss/index.js'
 export * from './params__ctx__memo_/index.js'
 export * from './signal__toggle_/index.js'